increase robustness

+++ src/main/java/io/bioimage/modelrunner/pytorch/tensor/ImgLib2Builder.java
@@ -113,7 +113,7 @@ public class ImgLib2Builder {
 				tensorShape);
 			byte val = flatArr[flatPos];
 			if (val < 0)
-				tensorCursor.get().set(256 + val);
+				tensorCursor.get().set(256 + (int) val);
 			else
 				tensorCursor.get().set(val);
 		}
